requirements.txt
asgiref==3.8.1
Django==4.2
django-braces==1.16.0
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
pillow==11.0.0
sqlparse==0.5.2
tzdata==2024.2
